[BMSPT-1376] refactoring parser -> converter
