+++ Tests/Parser/Xml21/Xml21Tests.cs
@@ -30,7 +30,7 @@ public class Xml21Tests {
   public async Task convertsBcfToJsonFiles21() {
     //var parser = new BcfParser<bcf21.Markup, bcf21.VisualizationInfo>();
     var json21 = new JsonConverter<bcf21.Markup, bcf21.Version>(jsonSerializer);
-    var markups = await BcfParser.ParseMarkups<bcf21.Markup, bcf21.VisualizationInfo>("Resources/TestBcf21.zip");
+    var markups = await BcfConverter.ParseMarkups<bcf21.Markup, bcf21.VisualizationInfo>("Resources/TestBcf21.zip");
     await json21.WriteMarkupsJson(markups, "Resources/output/21/json");
   }
   
@@ -38,14 +38,14 @@ public class Xml21Tests {
   public async Task convertsBcfMarkupsToJsonFiles30() {
     //var parser = new BcfParser<bcf30.Markup, bcf30.VisualizationInfo>();
     var json21 = new JsonConverter<bcf30.Markup, bcf30.Version>(jsonSerializer);
-    var markups = await BcfParser.ParseMarkups<bcf30.Markup,bcf30.VisualizationInfo>("Resources/TestBcf30.zip");
+    var markups = await BcfConverter.ParseMarkups<bcf30.Markup,bcf30.VisualizationInfo>("Resources/TestBcf30.zip");
     await json21.WriteMarkupsJson(markups, "Resources/output/30/json");
   }
   
   [Test]
   public async Task convertsBcfRootsToJsonFiles30() {
-    var extensions = await BcfParser.ParseExtensions<bcf30.Extensions>("Resources/TestBcf30.zip");
-    var project = await BcfParser.ParseProject<bcf30.ProjectInfo>("Resources/TestBcf30.zip");
+    var extensions = await BcfConverter.ParseExtensions<bcf30.Extensions>("Resources/TestBcf30.zip");
+    var project = await BcfConverter.ParseProject<bcf30.ProjectInfo>("Resources/TestBcf30.zip");
 
     var roots = new { Extensions = extensions, ProjectInfo = project };
     var json21 = new JsonConverter<bcf30.Markup, bcf30.Version>(jsonSerializer);
