+++ bcf-converter/Converter/Bcf/XDocumentExtensions.cs
@@ -1,11 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Xml.Linq;
 using System.Xml.Serialization;
 using bcf.bcf21;
 
-namespace bcf.Parser; 
+namespace bcf.Converter; 
 
 /// <summary>
 ///   A list of convenience extension methods on the XDocument class.
