+++ bcf-converter/Converter/Bcf/XElementExtensions.cs
@@ -2,7 +2,7 @@ using System;
 using System.Xml.Linq;
 using bcf.bcf21;
 
-namespace bcf.Parser;
+namespace bcf.Converter;
 /// <summary>
 ///   A list of convenience extension methods on the XElementExtensions class.
 ///   The BCF version specific parsing logic is defined here.
